Rename main module to tabmini
