+++ tabmini/estimators/HyperFast.py
